Split code further
